Separate Subject Set Component and Container
